Countermeasres parameters.  Lets users specify how easily missiles are spoofed by countermeasures
A missile is given a seeker strength.  Aspect default of 2.  Heat default of 3.  This is a missile's resistance to spoofage
A countermesaure is given an effectiveness for each type.  The defaults for countermeasure heat and aspect effectiveness are both 1.
Also gave the countermeasure an effective range.  Default is 300 meters.

A missiles chance of being spoofed is calculated by <cm type effectiveness>/<seeker strength>
So by default, 1/2 of aspect seekers will be fooled by a countermeasures, while 1/3 of heat seekers would be spoofed by default.

Also i gave the option of aspect missiles having a view cone different from 180 degrees.  This parameter is optional and doesn't effect anything already there.