Add "apply recoil" flag
This creates a flag that can be used in weapons.tbl to make a weapon
induce recoil effects.

If you don't multiply by numtimes, you only get the effect of firing one
shot even if you fire from multiple firing points at once or, when using
"$Shots:", you fire multiple rounds from a single firing point.

The reason you unrotate firing_vec at all is because it can't be
accessed outside of the convergence and autoaiming code above it unless
the firing_vec variable is moved out of the block, so you have to
redeclare it in the recoil block and unrotate it through a matrix to
even have a value to work with. As for why it's unrotated through
firing_orient, it's so that you get the recoil effect being applied
along the vector in which the shot is supposed to go, not along the
direction of travel of the ship, think of it as taking into account the
"gimbaled" nature of autoconverging and autoaiming weapons.