Make model_really_render()'s depth check use View_position instead of Eye_position.
View_position is modified by g3_start_instance_matrix()/g3_done_instance(). Eye_position is not. As a result, this depth check was basically comparing the distance from a small distance away from the center of the mission to the absolute coordinates of the camera, instead of their relative distances.