Use SDL_GetPrefPath for determining config directory
This tried to stay compatible with current installs as much as possible but for pilot files that isn't possible without rather substantial rewrites of the cfile system. Instead the launcher will handle copying the files from the old location.
These changes also will use the config directory for all writing operations on Windows similar to how Unix is currently handled.