Add generic configuration system
I would like to add an in-engine options screen at some point but to do
that we first need a system that is actually able to handle loading and
saving of various options values in a generic way.

These changes add a config system on top of the existing system that
provide a easy to use API for defining configuration options that are
automatically read and written to the config file at the appropriate
moments.