Don't use discard in main model shader
The only remaining occurence of discard was for when blending was
disabled and the pixel of the texture had an alpha value less than 0.95.
That combination cannot happen anymore since all transparent textures
are handled using one of the other blend modes and all other textures
won't have an alpha value less than 0.95.

This has some minor performance benefits in GPU limited scenes.