Check source code formatting with clang-format and Travis CI
This uses clang-format-diff for detecting improperly formatted lines on
the CI servers. The check is only applied to lines that were changed in
a pull request so this does not require a full source code reformatting
but still ensures that the style set by clang-format is used by new
commits.