replace most Errors in sexp.cpp with Warnings
This is a follow-up to commit 1614384a6ab35a5e63909dd825a6a6acb3d31aff.  `Error` is to be avoided if possible because almost all sexp problems are recoverable, and we don't want to bring the entire game to a halt.

Also add some more information to warnings about sexp operators on strings.