Aiturret Cleanup Take 2
While probably not a full fix to #1791 , this should clean up the portion of the code that looks the most suspicious, by making the checks for weapon flags for the whole block instead of just part.

All these changes are just changing the checking order, and should not affect the actual AI  I apologize in advance for any mistakes, since this ended up being much more involved than I thought.