Engine Wash Optimizations
A few optimizations to the Engine Wash function I found while bug hunting.

The engine was calculating the same value for tan and cos of ewp->angle.  These values never change, so setting them on table parse is trivial and far more optimal.  I think this function is supposed to be called 4 times a second for every ship, so I believe the optimization is worthwhile.

Secondly, on slower computers, engine wash was likely occassionally firing extra times because a check used less than instead of less than or equal to, forcing a weird situation in the timestamp function, making the function do the whole calculation again the next frame.

Third, pre-multiply a value that gets used more than once later on, so we don't square it every time we use it.

Last, we were calling the damage function for no reason when we weren't going to apply damage, anyway.

Very open to suggestions/corrections on this, since I am attempting to optimize.

*ALSO* Now simply skipping part of the loop if speed_scale = 0.0, since that precludes engine wash anyway.