add Smart_secondary_evade_chance
This adds a new AI profile flag that adds a new chance to evade
secondaries, that is *not* based on countermeasure availability.

Basing secondary evades off of CM count (done without this flag) means
the AI is really bad at evading secondaries in a bunch of situations.

One very notable example is that they're terrible at evading trebuchets
- without this flag, most trebuchet engagements (assuming the AI has
more than 8 countermeasures, so doesn't evade), are totally successful.

I've run through many repeated (I repeat each test 10 times) tests with
this flag, using an Ares and BP2-Captain as the AI profiles.

Without the flag set, or with it set to 0, the AI cannot dodge
trebuchets (0/10 times).

With the flag set to below 0.4 or so, the AI cannot dodge trebuchets
very well at all either (0/10 times).

Once the flag gets up to 0.5, the chance of dodging trebs increases
(5/10 times).

Once the flag gets up to 0.7, the chance of dodging trebs increases so
much that the AI nearly always dodges them (9/10 times are a dodge).

With a more agile ship, the evasion chance increases.