Fix overriding the default values set in the control table for disabled and continuous_ongoing
Previously, `control_config_common_init()` started with a loop that set the `disabled` and `continuous_ongoing` values to false for all controls in the config array. This overrides the default values set in the array. For example, if a control is set to be disabled by default then this loop overrides that incorrectly.

These lines were added in:

https://github.com/scp-fs2open/fs2open.github.com/commit/45da46eddd9d9249bf1178ac18164e4ab43570e1#diff-45a0618f2fa10c25b123d4a771f98228

and

https://github.com/scp-fs2open/fs2open.github.com/commit/61d9ce895be392c048437e985cf583ee2fc6161a#diff-45a0618f2fa10c25b123d4a771f98228

The goal of adding these lines was to initiate the values, but these values were already set in the struct table and those lines simply override the values. No other fields in the config array are set like this and simply use the values set in the array.

In tests I have removed these lines and printed what the default values are and the print statements print what is expected from the config array.

This PR removes those unneeded lines and thus allows the fields for `disabled` and `continuous_ongoing` to be respected as the default. Of course table changes can be used to alter these, but that's only if a controlconfigdefaults.tbl is present.